+++ lib/src/model/recording_interval.dart
@@ -5,4 +5,21 @@ enum RecordingInterval {
 
   /// 5 second interval
   interval5s;
+
+  /// Create an [RecordingInterval] from json
+  static RecordingInterval fromJson(dynamic json) {
+    switch (json as int) {
+      case 1:
+        return RecordingInterval.interval1s;
+      case 5:
+        return RecordingInterval.interval5s;
+      default:
+        throw Exception('Unknown RecordingInterval: $json');
+    }
+  }
+
+  /// Convert a [RecordingInterval] to json
+  dynamic toJson() {
+    return RecordingInterval.values.indexOf(this);
+  }
 }